+++ src/Service.Liquidity.Portfolio.Domain.Models/AssetBalanceDifference.cs
@@ -5,11 +5,11 @@ namespace Service.Liquidity.Portfolio.Domain.Models
         public string BrokerId { get; set; }
         public string WalletName { get; set; }
         public string Asset { get; set; }
-        public double Volume { get; set; }
-        public double VolumeUsd { get; set; }
-        public double CurrentPriceInUsd { get; set; }
+        public decimal Volume { get; set; }
+        public decimal VolumeUsd { get; set; }
+        public decimal CurrentPriceInUsd { get; set; }
 
-        public AssetBalanceDifference(string brokerId, string walletName, string asset, double volume, double volumeUsd, double currentPriceInUsd)
+        public AssetBalanceDifference(string brokerId, string walletName, string asset, decimal volume, decimal volumeUsd, decimal currentPriceInUsd)
         {
             BrokerId = brokerId;
             WalletName = walletName;
