+++ src/Service.Liquidity.Portfolio.Domain/Models/NetBalanceByAsset.cs
@@ -8,8 +8,8 @@ namespace Service.Liquidity.Portfolio.Domain.Models
     {
         [DataMember(Order = 1)] public string Asset { get; set; }
         [DataMember(Order = 2)] public List<NetBalanceByWallet> WalletBalances { get; set; }
-        [DataMember(Order = 4)] public double NetVolume { get; set; }
-        [DataMember(Order = 5)] public double NetUsdVolume { get; set; }
+        [DataMember(Order = 4)] public decimal NetVolume { get; set; }
+        [DataMember(Order = 5)] public decimal NetUsdVolume { get; set; }
         
         public NetBalanceByAsset()
         {
